{{ cookiecutter.app_name }}/{{ cookiecutter.app_name }}/requirements/dev.txt
-r prod.txt

# Debug Toolbar
Flask-DebugToolbar==0.9.0
